requirements.txt
pydub
librosa
noisereduce
google-genai
requests
filetype
torchaudio==2.7.1
speechbrain==1.1.0
hyperpyyaml
julius
fastapi
uvicorn
python-dotenv
portalocker
numpy
soundfile
pedalboard
python-multipart
